unify json generator and triple generator in one file
